requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aiosignal==1.3.1
albucore==0.0.20
albumentations==1.4.21
annotated-types==0.7.0
APScheduler==3.11.0
apscheduler-di==0.1.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
black==24.10.0
certifi==2024.7.4
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
easydict==1.13
eval_type_backport==0.2.0
fonttools==4.55.0
frozenlist==1.4.1
idna==3.7
imageio==2.36.1
insightface==0.7.3
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
magic-filter==1.0.12
matplotlib==3.9.2
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.1.3
onnx==1.17.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
prettytable==3.12.0
protobuf==5.29.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
rodi==2.0.6
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
simsimd==6.2.1
six==1.16.0
stringzilla==3.10.10
threadpoolctl==3.5.0
tifffile==2024.9.20
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
wcwidth==0.2.13
yarl==1.9.4
